+++ CHANGELOG.md
@@ -5,7 +5,12 @@ SPDX-License-Identifier: LicenseRef-GPL-2.0-only-with-DUNE-exception
 
 # Master (will become release 2.12)
 
-- ...
+- Add `LagrangeSimplexLocalFiniteElement` with dynamic order, by generalizing the existing
+  static order implementation. The polynomial order of the shape functions is passed as constructor
+  parameter instead of a template parameter (which is implicitly set to `-1` as an indicator for the
+  dynamic polynomial order). The dynamic implementation uses internally a mutable buffer during the
+  shape function evaluation, which makes its usage potentially unsafe when an instance is shared
+  between concurrent threads.
 
 # Release 2.11